+++ src/__tests__/periodicaorder.test.js
@@ -1,19 +1,5 @@
-import { graphql } from "graphql";
-import { getExecutableSchema } from "../schemaLoader";
 import { createMockedDataLoaders } from "../datasourceLoader";
-
-export async function performTestQuery({ query, variables, context }) {
-  return graphql(
-    await getExecutableSchema({
-      loadExternal: false,
-      clientPermissions: { admin: true },
-    }),
-    query,
-    null,
-    context,
-    variables
-  );
-}
+import { performTestQuery } from "../utils/utils";
 
 const query = `
 mutation($input: PeriodicaArticleOrder!) {
