+++ src/__tests__/search.test.js
@@ -1,23 +1,8 @@
 /**
  * @file This file contains tests for the search operation
  */
-
-import { graphql } from "graphql";
-import { getExecutableSchema } from "../schemaLoader";
 import { createMockedDataLoaders } from "../datasourceLoader";
-
-export async function performTestQuery({ query, variables, context }) {
-  return graphql(
-    await getExecutableSchema({
-      loadExternal: false,
-      clientPermissions: { admin: true },
-    }),
-    query,
-    null,
-    context,
-    variables
-  );
-}
+import { performTestQuery } from "../utils/utils";
 
 test("search - with all filters and facets", async () => {
   const result = await performTestQuery({
