+++ src/__tests__/submitOrder.test.js
@@ -3,22 +3,8 @@
  *
  */
 
-import { graphql } from "graphql";
-import { getExecutableSchema } from "../schemaLoader";
 import { createMockedDataLoaders } from "../datasourceLoader";
-
-export async function performTestQuery({ query, variables, context }) {
-  return graphql(
-    await getExecutableSchema({
-      loadExternal: false,
-      clientPermissions: { admin: true },
-    }),
-    query,
-    null,
-    context,
-    variables
-  );
-}
+import { performTestQuery } from "../utils/utils";
 
 test("submitorder fails when user is not authenticated, and no userId provided", async () => {
   const result = await performTestQuery({
